perms: move perms.go into poolrpc
This change narrows the import scope for lightning-node-connect (LNC).
Instead of importing the entire pool package, LNC now only needs to
import poolrpc.

It also brings pool in line with the structure used in our other
projects.
